+++ packages/next/client/components/hooks-client-context.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { createContext } from 'react'
 
 export const SearchParamsContext = createContext<URLSearchParams>(null as any)
