+++ packages/next/shared/lib/app-router-context.ts
@@ -1,3 +1,6 @@
+'use client'
+'use client'
+
 import React from 'react'
 import type { FocusAndScrollRef } from '../../client/components/reducer'
 import type { FlightRouterState, FlightData } from '../../server/app-render'
